requirements.txt
django==2.2.6
django_cleanup
psycopg2
dj_database_url
django-crispy-forms
geonamescache
pillow
